pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "guardintent"
version = "0.1.0"
description = "CLI security automation and triage framework"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "GuardIntent"}]
dependencies = [
  "typer>=0.12.0",
  "rich>=13.7.0",
  "pyyaml>=6.0.1",
]

[project.scripts]
guardintent = "guard_intent.cli:app"

[tool.setuptools.packages.find]
include = ["guard_intent*"]
exclude = ["data*", "reports*", "tests*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

